+++ include/cpp3ds/actors/Rectangle.h
@@ -13,9 +13,9 @@ namespace cpp3ds {
 			int width, height;
 		public:
 			Color color;
-			Rectangle(int w, int h, Color color, int x = 0, int y = 0, int depth3d = 0):
+			Rectangle(int w, int h, Color color, float x = 0, float y = 0, float depth3d = 0):
 			        width(w), height(h), color(color), Actor(x, y, depth3d) {}
-			virtual void draw(Screen& screen, int x = 0, int y = 0);
+			virtual void draw(Screen& screen, float x = 0, float y = 0);
 		};
 
 	}
